consolidate all helpers into common.go
